Match Data Is Fetchable
What i just implemented is a page that shows my api pathing to get match data from an api, but specifically within a certain rank. It also was a page i used for testing to see what was wrong or right about my code. It will not be kept in the project as a live page, but for now that is the best way for me to be able to use data.